% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{easy_boxplot}
\alias{easy_boxplot}
\title{Create a Boxplot}
\usage{
easy_boxplot(dataset, x, y, color, ...)
}
\description{
Uses inserted variables and data set to create a boxplot that will be displayed in the RStudio plot window.
}
