% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{easy_histogram}
\alias{easy_histogram}
\title{Quickly Create a Histogram}
\usage{
easy_histogram(dataset, x, fill = NULL, ...)
}
\arguments{
\item{dataset}{The data frame providing the data for the plot.}

\item{x}{The variable name to be used for the x-axis.}

\item{fill}{A categorical variable to fill the histogram.}

\item{...}{Additional arguments passed to \code{geom_histogram()}, such as \code{size}, \code{alpha}, etc.}
}
\value{
A ggplot2 histogram object.
}
\description{
Given one numeric variable from a data frame, the \code{easy_histogram()} function provides a simpler
way to create histograms without manually writing full ggplot2 code.
Users can customize the setting of the histogram by adding additional arguments.
}
\examples{
easy_histogram(mtcars, x = "mpg", fill = "cyl")
}
