% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{easy_lineplot}
\alias{easy_lineplot}
\title{Quickly Create a Lineplot}
\usage{
easy_lineplot(dataset, x, y, color = NULL, ...)
}
\arguments{
\item{dataset}{The name of the data frame providing the data for the plot.}

\item{x}{The variable name to be used for the x-axis (the independent variable).}

\item{y}{The variable name to be used for the y-axis (the dependent variable).}

\item{color}{A categorical variable to be used to color the plot.}

\item{...}{Additional ggplot2 arguments passed to \code{geom_point()}, such as \code{size}, \code{alpha}, etc.}
}
\value{
A ggplot2 lineplot displayed in Rstudio plot window.
}
\description{
Given two numeric variables, the \code{easy_lineplot()} function provides a simpler way to display a lineplot in the Rstudio plot window without manually writing the full ggplot2 code.
Users can customize the setting of the lineplot by adding additional arguments.
}
\examples{
easy_lineplot(mtcars, x = "hp", y = "mpg", color = "gear")
}
