% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{easy_scatterplot}
\alias{easy_scatterplot}
\title{Quickly Create a Scatterplot}
\usage{
easy_scatterplot(dataset, x, y, color = NULL, ...)
}
\arguments{
\item{dataset}{The data frame providing the data for the plot.}

\item{x}{The variable name to be used for the x-axis (the independent variable).}

\item{y}{The variable name to be used for the y-axis (the dependent variable).}

\item{color}{A categorical variable to color the points.}

\item{...}{Additional ggplot2 arguments passed to \code{geom_point()}, such as \code{size}, \code{alpha}, etc.}
}
\value{
A ggplot2 scatterplot object.
}
\description{
Given two numeric variables from a data frame, the \code{easy_scatterplot()} function provides a simpler
way to create scatterplots without manually writing full ggplot2 code.
Users can customize the setting of the scatterplot by adding additional arguments.
}
\examples{
easy_scatterplot(mtcars, x = "wt", y = "mpg", color = "cyl", size = 3, alpha = 0.5)
}
