% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{plot_helper}
\alias{plot_helper}
\title{Plot Helper}
\usage{
plot_helper(dataset, x, y)
}
\arguments{
\item{dataset}{The name of the data frame providing the data for the plot.}

\item{x}{The variable name to be used for the x-axis (the independent variable).}

\item{y}{The variable name to be used for the y-axis (the dependent variable).}
}
\value{
A character string.
}
\description{
Uses inserted dataset, x-value, and y-value to return a suggestion on what type of plot the user should make with the inserted variable types.
}
\examples{
plot_helper(mpg, x = "cty", y = "hwy")
}
