% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottie.R
\name{easy_boxplot}
\alias{easy_boxplot}
\title{Quickly Create a Boxplot}
\usage{
easy_boxplot(dataset, x, y, color, ...)
}
\arguments{
\item{dataset}{The name of the data frame providing the data for the plot.}

\item{x}{The variable name to be used for the x-axis (the independent variable).}

\item{y}{The variable name to be used for the y-axis (the dependent variable).}

\item{color}{A categorical variable to be used to color the plot.}

\item{...}{Additional ggplot2 arguments passed to \code{geom_point()}, such as \code{size}, \code{alpha}, etc.}
}
\value{
A ggplot2 boxplot displayed in Rstudio plot window.
}
\description{
Given one discrete variable and one numeric variable, the \code{easy_boxplot()} function provides a simpler way to display a boxplot in the Rstudio plot window without manually writing the full ggplot2 code.
Users can customize the setting of the boxplot by adding additional arguments.
}
\examples{
easy_boxplot(mtcars, x = "cyl", y = "mpg", color = "gear")
}
